add pay debt fraction button handler + some other fixes

+++ trunk/stockManagement/src/com/pfe/server/service/ShipmentServiceImpl.java
@@ -20,9 +20,9 @@ import com.pfe.shared.model.ProductType;
 import com.pfe.shared.model.Shipment;
 import com.pfe.shared.model.Stock;
 
-@Service("shipmentService") 
+@Service("shipmentService")
 public class ShipmentServiceImpl implements ShipmentService {
-	
+
 	@Autowired
 	private ShipmentDao shipmentDao;
 	@Autowired
@@ -34,48 +34,54 @@ public class ShipmentServiceImpl implements ShipmentService {
 	@Autowired
 	private DozerBeanMapper dozerMapper;
 
-
 	@Override
-	public void deleteList(List<ShipmentDTO> shipments) throws BusinessException {
-		
+	public void deleteList(List<ShipmentDTO> shipments)
+			throws BusinessException {
+
 		List<Shipment> entities = new ArrayList<Shipment>();
-		for(ShipmentDTO shipment : shipments){
+		for (ShipmentDTO shipment : shipments) {
 			entities.add(dozerMapper.map(shipment, Shipment.class));
 		}
 
 		//Manage exceptions
-		for(Shipment entity : entities){
-			
-			//if part of the shipment was sold
-			if(entity.getCurrentQuantity() != entity.getInitialQuantity()){
-				throw new BusinessException("Some of the shipments are impossible to delete because" +
-						" items have been sold.");
-			}
-			
-			//if not enough goods in warehouse 
-			LocationType warehouseType = locationTypeDao.getWarehouseType();
-			Location warehouse = locationDao.get(warehouseType).get(0);	
-			ProductType type = entity.getProductType();
-			Stock stock = stockDao.get(warehouse, type);
-			if(stock == null){
-				throw new BusinessException("Some of the shipments are impossible to delete because" +
-						" there are not enought goods in the warehouse");
-			} else{
-				int availableQty = stock.getQuantity();
-				int shipmentQty = entity.getInitialQuantity();
-				if(availableQty < shipmentQty){
-					throw new BusinessException("Some of the shipments are impossible to delete because" +
-							" there are not enought goods in the warehouse");
+		for (Shipment entity : entities) {
+
+			//If shipment is in the database
+			if (entity.getId() != null) {
+				
+				//Part of the shipment was sold
+				if (entity.getCurrentQuantity() != entity.getInitialQuantity()) {
+					throw new BusinessException(
+							"Some of the shipments are impossible to delete because"
+									+ " items have been sold.");
 				}
-				else{
-					//delete shipment and update warehouse stocks
-					int remainingQty = availableQty - shipmentQty;
-					stock.setQuantity(remainingQty);
-					stockDao.merge(stock);
-					shipmentDao.delete(entity);
+
+				//Not enough goods in warehouse
+				LocationType warehouseType = locationTypeDao.getWarehouseType();
+				Location warehouse = locationDao.get(warehouseType).get(0);
+				ProductType type = entity.getProductType();
+				Stock stock = stockDao.get(warehouse, type);
+				if (stock == null) {
+					throw new BusinessException(
+							"Some of the shipments are impossible to delete because"
+									+ " there are not enought goods in the warehouse");
+				} else {
+					int availableQty = stock.getQuantity();
+					int shipmentQty = entity.getInitialQuantity();
+					if (availableQty < shipmentQty) {
+						throw new BusinessException(
+								"Some of the shipments are impossible to delete because"
+										+ " there are not enought goods in the warehouse");
+					} else {
+						
+						//Delete shipment and update warehouse stocks
+						int remainingQty = availableQty - shipmentQty;
+						stock.setQuantity(remainingQty);
+						stockDao.merge(stock);
+						shipmentDao.delete(entity);
+					}
 				}
 			}
 		}
-		
 	}
 }
